PSD-976: Commonize document visibility logic
