+++ psd-web/app/helpers/documents_helper.rb
@@ -69,20 +69,4 @@ module DocumentsHelper
       "Updated #{Time.zone.parse(file.blob.metadata[:updated]).strftime('%d/%m/%Y')}"
     end
   end
-
-  def document_is_sensitive(document, parent)
-    document_activity = document.attachments.find_by(record_type: "Activity")&.record
-    child_is_visible = document_activity&.source&.user_has_gdpr_access? || !document.metadata[:has_consumer_info]
-    parent_forces_to_be_visible = parent.respond_to?(:child_should_be_displayed?) && parent.child_should_be_displayed?
-    !(child_is_visible || parent_forces_to_be_visible)
-  end
-
-  def document_sensitive_title
-    "Attachment restricted"
-  end
-
-  def document_sensitive_body
-    "This attachment is restricted because it has been marked as containing GDPR protected data. " +
-      "Contact the case owner if you need access."
-  end
 end
